Implement NetworkRequest, images, doc strings
NetworkRequest is great in that it makes the UI pretty responsive
while getting documents from the service, but has some drawbacks.
It's hard to the proxy model to sort correctly, at least initially,
since the objects may not have their document yet, and so no title,
so nothing to sort against.

Also, the fetchMore logic somehow makes the app crash. :(