Add request "batches", some tweaks
Previously, the list view would continue showing empty rows
(rows for which there is no document) as well as a random sort
order  until widgets were interacted with,
like when clicking the image-only button, or changing the
sort order.

We now have a mechanism to keep track of how many requests have
completed vs how many we sent out during a cycle of queue processing.
When all the requests in a batch have completed, the proxy model
is asked to refresh, causing the empty rows to disappear and the
items to be sorted correctly.

In addition, some minor tweaks like doc strings.