[BUG] Fix longer-than-expected empty lines.
After implementing batching of requests, there was a blind spot
remaining: get_object(). This was called in two places:
- When processing the queue;
- When calling model.data().

The first use was going through the batching process.
The queue would then be waiting for 60 seconds to process the next
batch of requests.

But model.data() was bypassing it for new requests. Because the qeueue
was waiting, the requests were processed but the post-process logic
happening after batch wasn't, resulting in rows remaining empty
for longer than expected (perhaps permanently).

To avoid confusion, removed the ObjectCache.objects property
accessors, so that the only way to access the object documents
is the get_objects() method.